Save Progress

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -106,7 +106,7 @@ return new class extends Migration
             $table->string('ip_address', 45)->nullable();
             $table->text('user_agent')->nullable();
             $table->longText('payload');
-            $table->integer('last_activity')->index();
+            $table->unsignedBigInteger('last_activity')->index();
             $table->jsonb('metadata')->default(json_encode([
                 'created_at' => null,
                 'created_by' => null,
